+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- **[Changed]** streamlined `ParallelState` to prevent potential interrupt race issues. Enable `AtLeastOneState` to overwrite two function for its functionality.
 - **[Fixed]** condition in `AtLeastOneState` where other states aren't destroyed when exiting under success condition.
 
 ## [0.3.5] - 2021-06-15